refactor(node): Split out code to detect the right Node package manager
Reduce the `NpmSupport` file length for a better overview.
